requirements.txt
aiofiles==0.8.0
aiohttp==3.8.6
aioredis==1.3.1
aiosignal==1.3.1
alembic==1.11.1
amqp==5.2.0
anyio==4.4.0
api-analytics==1.2.3
# Editable install with no version control (arbm_core==0.4.1)
# -e C:\Users\HomePC\Downloads
-e git+https://github.com/Anita-Mutemi/arbm-core.git@main#egg=arbm-core


async-timeout==4.0.3
attrs==23.2.0
autoflake==1.5.3
bandit==1.7.9
bcrypt==4.0.1
billiard==4.2.0
black==22.6.0
cached_classproperty==1.0.1
celery==5.3.6
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chart-studio==1.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.5.4
cryptography==42.0.8
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.0.0.post2
fastapi==0.95.2
fastapi-cache==0.1.0
fastapi-cache2==0.2.1
freezegun==1.2.2
frozenlist==1.4.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isort==5.10.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
linkedin-api==1.1.0
loguru==0.7.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
notifiers==1.3.3
numpy==1.25.1
openai==0.27.10
ordered_enum==0.0.8
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pandas==2.0.2
passlib==1.7.4
pathspec==0.12.1
pbr==6.0.0
pendulum==3.0.0
Pillow==9.5.0
platformdirs==4.2.2
plotly==5.10.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.6
pyasn1==0.6.0
pycparser==2.22
pydantic==1.10.11
pydantic-duality==1.1.1
pyflakes==3.2.0
Pygments==2.18.0
pymongo==4.8.0
PySocks==1.7.1
pytest==7.3.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2023.3
pyupgrade==2.37.3
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.14.0
selenium==4.4.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.49
starlette==0.27.0
stevedore==5.2.0
tenacity==8.5.0
threadpoolctl==3.5.0
time-machine==2.14.2
tokenize-rt==5.2.0
toml==0.10.2
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.30.1
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
